\name{algoio.publish}
\alias{algoio.publish}
\title{Publish entire R package to Algorithms.io platform}
\usage{
  algoio.publish(package.dir)
}
\arguments{
  \item{package.dir}{the package's top directory}
}
\value{
  response from Algorithms.io platform including endpoint
  of the new package
}
\description{
  Publish entire R package to Algorithms.io platform
}
\author{
  anthony@algorithms.io
}
\keyword{publish}
\keyword{upload,}

